dev10 changes

+++ src/ansys/meshing/prime/numen/surface.py
@@ -20,8 +20,6 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-"""Module for surface mesh."""
-
 import ansys.meshing.prime.numen.controls as controls
 import ansys.meshing.prime.numen.utils.macros as macros
 from ansys.meshing import prime
@@ -31,10 +29,10 @@ from ansys.meshing.prime.internals.error_handling import (
 )
 from ansys.meshing.prime.numen.utils import surface_utils
 from ansys.meshing.prime.numen.utils.cached_data import CachedData, create_size_control
+from ansys.meshing.prime.numen.utils.communicator import call_method
 
 
 def mesh(model: prime.Model, mesh_params: dict, cached_data: CachedData):
-    """Perform surface mesh based on sizing type."""
     size_field_names = mesh_params["size_field_names"]
     part_scope = mesh_params["part_scope"]
     part_ids = macros._get_part_ids(model, part_scope)
@@ -68,10 +66,10 @@ def mesh(model: prime.Model, mesh_params: dict, cached_data: CachedData):
 
 
 def mesh_with_sizing(model: prime.Model, mesh_params: dict, cached_data: CachedData):
-    """Perform surface mesh based on sizing."""
     part_scope = mesh_params["part_scope"]
     retain_generated_size_field = mesh_params["retain_generated_size_field"]
     sf_per_part = mesh_params["size_field_computation_per_part"]
+    rbf = mesh_params["rbf_interpolation"]
     model.set_global_sizing_params(
         prime.GlobalSizingParams(
             model, mesh_params["min_size"], mesh_params["max_size"], mesh_params["growth_rate"]
@@ -81,8 +79,9 @@ def mesh_with_sizing(model: prime.Model, mesh_params: dict, cached_data: CachedD
 
     size_field = prime.SizeField(model)
     periodic_params = prime.SFPeriodicParams(model, angle=0.0)
+    threaded_sf = mesh_params["_threaded_size_field_compute"]
     size_field_params = prime.VolumetricSizeFieldComputeParams(
-        model, periodic_params=periodic_params
+        model, periodic_params=periodic_params, enable_multi_threading=threaded_sf
     )
     size_controls = mesh_params["size_controls"]
     if not sf_per_part:
@@ -98,21 +97,30 @@ def mesh_with_sizing(model: prime.Model, mesh_params: dict, cached_data: CachedD
 
         size_field_res = size_field.compute_volumetric(size_control_ids, size_field_params)
 
+    threaded_surfer = mesh_params["_threaded_surfer"]
     surfer_params = prime.SurferParams(
         model=model,
         size_field_type=prime.SizeFieldType.VOLUMETRIC,
         project_on_geometry=True,
-        enable_multi_threading=False,
+        enable_multi_threading=threaded_surfer,
+        enableRobustMode=rbf,
+        retryWithRBFInterpolation=rbf,
     )
     surfer = prime.Surfer(model)
     for part_id in part_ids:
         part = model.get_part(part_id)
-        print("inprogress=======================\n")
-        print(part.name)
         if sf_per_part:
             size_control_ids = []
             for size_control in size_controls:
-                if macros._check_size_control_scope(
+                if size_control["control_type"] == 'boi':
+                    size_control_params = {
+                        "sizing_type": controls._get_size_control_type(size_control),
+                        "sizing_params": controls._get_size_control_params(model, size_control),
+                        "scope": controls._get_scope(model, size_control),
+                    }
+                    size_control_id = create_size_control(model, size_control_params, cached_data)
+                    size_control_ids.append(size_control_id)
+                elif macros._check_size_control_scope(
                     model,
                     part_id,
                     size_control["entity_type"],
@@ -134,6 +142,9 @@ def mesh_with_sizing(model: prime.Model, mesh_params: dict, cached_data: CachedD
                         )
                         size_control_ids.append(size_control_id)
             size_field_res = size_field.compute_volumetric(size_control_ids, size_field_params)
+            size_field_name = mesh_params["size_field_per_part_prefix"] + "_" + part.name
+            args = {"size_field_id": size_field_res.size_field_id, "name": size_field_name}
+            call_method(model, "PrimeMesh::Model/SetSizeFieldName", model._object_id, args)
         part = model.get_part(part_id)
         topo_faces = part.get_topo_faces()
         if len(topo_faces) > 0:
@@ -153,10 +164,15 @@ def mesh_with_sizing(model: prime.Model, mesh_params: dict, cached_data: CachedD
                         for face in topo_faces
                         if not model.topo_data.get_mesh_zonelets_of_topo_faces([face])
                     ]
-                    print("Unmeshed topo face ids " + str(unmeshed_faces) + " on part " + part.name)
+                    cached_data._logger.warning(
+                        "    Unmeshed topo face ids "
+                        + str(unmeshed_faces)
+                        + " on part "
+                        + part.name
+                    )
             except PrimeRuntimeError as error:
-                print(
-                    "Error with code '{}' on '{}' with message '{}'".format(
+                cached_data._logger.warning(
+                    "    Error with code '{}' on '{}' with message '{}'".format(
                         error.error_code,
                         part.name,
                         error.message,
@@ -165,27 +181,28 @@ def mesh_with_sizing(model: prime.Model, mesh_params: dict, cached_data: CachedD
                 if mesh_params["stop_on_failure"] == True:
                     break
             except PrimeRuntimeWarning as warning:
-                print(
-                    "Warning: '{}' with message '{}'.".format(
+                cached_data._logger.warning(
+                    "    Warning: '{}' with message '{}'.".format(
                         part.name,
                         warning.message,
                     )
                 )
         else:
-            cached_data._logger.warning(f"No topoface was found in part \"{part.name}\"")
-        if sf_per_part:
+            cached_data._logger.warning(f"    No topoface was found in part \"{part.name}\"")
+        if sf_per_part and retain_generated_size_field:
+            model.deactivate_volumetric_size_fields([size_field_res.size_field_id])
+        if not retain_generated_size_field:
             model.delete_volumetric_size_fields([size_field_res.size_field_id])
             model.control_data.delete_controls(size_control_ids)
             size_control_ids = []
-            retain_generated_size_field = True
 
-    if not retain_generated_size_field:
+    if not retain_generated_size_field and not sf_per_part:
         model.delete_volumetric_size_fields([size_field_res.size_field_id])
-    model.control_data.delete_controls(size_control_ids)
+    if len(size_control_ids) > 0:
+        model.control_data.delete_controls(size_control_ids)
 
 
 def mesh_with_size_controls(model: prime.Model, mesh_params: dict, cached_data: CachedData):
-    """Perform surface mesh with the given size control."""
     part_scope = mesh_params["part_scope"]
     part_ids = macros._get_part_ids(model, part_scope)
 
@@ -214,8 +231,7 @@ def mesh_with_size_controls(model: prime.Model, mesh_params: dict, cached_data: 
         cached_data.destroy_cached_object(size_control)
 
 
-def improve_quality(model: prime.Model, improve_quality_params: dict, cached_data: CachedData):
-    """Improve surface mesh quality."""
+def post_mesh_cleanup(model: prime.Model, improve_quality_params: dict, cached_data: CachedData):
     part_scope = improve_quality_params["part_scope"]
     soft_target_skewness = improve_quality_params["soft_target_skewness"]
     hard_target_skewness = improve_quality_params["hard_target_skewness"]
@@ -244,12 +260,30 @@ def improve_quality(model: prime.Model, improve_quality_params: dict, cached_dat
             local_remesh_by_size_change,
             keep_small_free_surfaces,
         )
+        diag_params = prime.SurfaceDiagnosticSummaryParams(
+            model,
+            scope=prime.ScopeDefinition(model, part_expression=part.name),
+            compute_self_intersections=True,
+            compute_free_edges=True,
+            compute_multi_edges=True,
+            compute_duplicate_faces=True,
+        )
+        surface_search_tool = prime.SurfaceSearch(model=model)
+        checks = surface_search_tool.get_surface_diagnostic_summary(diag_params)
+        if (
+            checks.n_free_edges > 0
+            or checks.n_multi_edges > 0
+            or checks.n_self_intersections > 0
+            or checks.n_duplicate_faces > 0
+        ):
+            info1 = f"{part.name} : {checks.n_free_edges} {checks.n_multi_edges}"
+            info2 = f" {checks.n_self_intersections} {checks.n_duplicate_faces}"
+            cached_data._logger.info(info1 + info2)
 
 
 def clean_up_triangles(
     model: prime.Model, clean_up_triangles_params: dict, cached_data: CachedData
 ):
-    """Clean up triangular elements."""
     part_scope = clean_up_triangles_params["part_scope"]
     collapse_sliver_faces = True
     stitch_free_faces = True
@@ -360,10 +394,6 @@ def clean_up_triangles(
             or checks.n_self_intersections > 0
             or checks.n_duplicate_faces > 0
         ):
-            print(
-                part.name,
-                checks.n_free_edges,
-                checks.n_multi_edges,
-                checks.n_self_intersections,
-                checks.n_duplicate_faces,
-            )
+            info1 = f"{part.name} : {checks.n_free_edges} {checks.n_multi_edges}"
+            info2 = f" {checks.n_self_intersections} {checks.n_duplicate_faces}"
+            cached_data._logger.info(info1 + info2)
